[#26898] 32bits compilation error : change declaration real to real(kind=8) in proten.F90/proten.h
Former-commit-id: ed986f0bcc60ce0752f3f3aea98a7150fdf9a9d9

+++ bibfor/algorith/proten.F90
@@ -18,8 +18,8 @@
 
 function proten(u,v) result(w)
     implicit none
-    real,dimension(:),intent(in) :: u,v
-    real,dimension(size(u),size(v)) :: w
+    real(kind=8),dimension(:),intent(in) :: u,v
+    real(kind=8),dimension(size(u),size(v)) :: w
     integer :: i,j
     do i =1, size(u)
         do j =1, size(v)
